perf(fm-core): shrink Position line/col/byte usize->u32 -- Span 48->24B, IrNode 208->184B, IrEdge 120->96B, parse bytes -10.8%
The last large IR field. Span (=2x Position, each 3x usize = 48B) sits on EVERY
IrNode (23% of 208B) and IrEdge (40% of 120B) but is only source-location for
diagnostics/source-mapping, never geometry -- no Mermaid source nears 4 GiB, so
Position's three fields drop usize->u32, halving Position (24->12B) and Span
(48->24B).

MEASURED (deterministic counting-allocator + size_of, load-immune):
  size_of Position 24->12B, Span 48->24B, IrNode 208->184B (-11.5%),
  IrEdge 120->96B (-20%); parse bytes wide_16x32 979,760 -> 873,728 (-10.8%).
  Alloc COUNT unchanged (1599) -- pure memory-traffic reduction, no new box.

Compiler-guided ~10 cast sites: construction casts internally via
u32::try_from(..).unwrap_or(u32::MAX); diagnostic reads widen `as usize`.
Position/Span are workspace-internal (fm-core has no crates.io consumers).

Byte-identical & serde-safe: u32/usize serialize identically as JSON numbers;
spans never affect layout geometry or rendered SVG. golden_svg (1),
golden_layout (2), frankentui_conformance (2) GREEN, NO re-bless; fm-core 350 +
fm-parser 405 + fm-render-svg 234 + fm-render-term 78 lib tests GREEN.

Also fixed pre-existing test-only breakage from 7b01bbe + bef271d never
recompiled: conformance + fm-render-svg + fm-parser test modules read the moved
icon/href/callback/tooltip fields as fields -> converted to () accessors /
interaction_mut(). Conformance gate is GREEN again.

Standing dominance (context, not a new claim): full_pipeline_wide vs pinned
live-CDP Mermaid 11.12.0 = 0.004280/0.003248/0.002136 at 8x16/12x24/16x32
(Mermaid 233.6x/307.9x/468.2x slower).

Session+prior struct-shrink total: IrNode 608->184B (-70%), IrEdge 256->96B
(-62.5%). Cheap struct-shrink frontier essentially closed.

+++ crates/fm-cli/src/main.rs
@@ -3376,7 +3376,7 @@ fn collect_fnx_diagnostics(results: &FnxAnalysisResults) -> Vec<ValidationDiagno
 
         let (source_line, source_column) = diag
             .span
-            .map(|span| (Some(span.start.line), Some(span.start.col)))
+            .map(|span| (Some(span.start.line as usize), Some(span.start.col as usize)))
             .unwrap_or((None, None));
 
         diagnostics.push(ValidationDiagnostic {
